I want to kill myself

+++ API/src/main/java/com/maimai/tamagotchi/item/DefaultType.java
@@ -6,5 +6,5 @@ public interface DefaultType {
 
     double getValue();
 
-    int getCost();
+    double getCost();
 }
