+++ API/src/main/java/com/maimai/tamagotchi/item/impl/FoodType.java
@@ -15,9 +15,9 @@ public enum FoodType implements DefaultType {
 
     private final String name;
     private final double value;
-    private final int cost;
+    private final double cost;
 
-    FoodType(String name, double value, int cost) {
+    FoodType(String name, double value, double cost) {
         this.name = name;
         this.value = value;
         this.cost = cost;
@@ -34,7 +34,7 @@ public enum FoodType implements DefaultType {
     }
 
     @Override
-    public int getCost() {
+    public double getCost() {
         return cost;
     }
 }
